Also organize compile command
