+++ inlang/source-code/paraglide/paraglide-js/src/cli/commands/compile/command.test.ts
@@ -3,11 +3,11 @@ import memfs from "memfs"
 import nodeFsPromises from "node:fs/promises"
 import { type NodeishFilesystem, createNodeishMemoryFs } from "@lix-js/fs"
 import { createMessage } from "@inlang/sdk/test-utilities"
-import { compileCommand } from "./compile.js"
+import { compileCommand } from "./command.js"
 import type { ProjectSettings } from "@inlang/sdk"
 import { resolve } from "node:path"
-import { Logger } from "../../services/logger/index.js"
-import { pathExists } from "../../services/file-handling/exists.js"
+import { Logger } from "../../../services/logger/index.js"
+import { pathExists } from "../../../services/file-handling/exists.js"
 
 beforeEach(() => {
 	vi.resetAllMocks()
